Move more functions to service layer
